--- docs/source/smoke_zephyr.memory_configuration.rst
@@ -1,21 +1,0 @@
-:mod:`smoke_zephyr.memory_configuration` --- General configuration manager
-===================================================================
-
-.. module:: smoke_zephyr.memory_configuration
-   :synopsis: General configuration manager
-
-Data
-----
-
-.. autodata:: smoke_zephyr.memory_configuration.SERIALIZER_DRIVERS
-   :annotation:
-
-.. autodata:: smoke_zephyr.memory_configuration.has_yaml
-
-Classes
--------
-
-.. autoclass:: smoke_zephyr.memory_configuration.MemoryConfiguration
-   :members:
-   :special-members: __init__
-   :undoc-members:
