+++ smoke_zephyr/configuration.py
@@ -49,8 +49,8 @@ else:
 
 SERIALIZER_DRIVERS = {}
 """The serializer drivers that are available."""
-SERIALIZER_DRIVERS['json'] = {'load': json.load, 'dumps': lambda obj: json.dumps(obj, sort_keys=True, indent=4)}
-SERIALIZER_DRIVERS['jsn'] = {'load': json.load, 'dumps': lambda obj: json.dumps(obj, sort_keys=True, indent=4)}
+SERIALIZER_DRIVERS['json'] = {'load': json.load, 'loads': json.loads, 'dumps': lambda obj: json.dumps(obj, sort_keys=True, indent=4)}
+SERIALIZER_DRIVERS['jsn'] = {'load': json.load, 'loads': json.loads, 'dumps': lambda obj: json.dumps(obj, sort_keys=True, indent=4)}
 if has_yaml:
 	SERIALIZER_DRIVERS['yaml'] = {'load': lambda file_obj: yaml.load(file_obj, Loader=Loader), 'dumps': lambda obj: yaml.dumps(obj, default_flow_style=False, Dumper=Dumper)}
 	SERIALIZER_DRIVERS['yml'] = {'load': lambda file_obj: yaml.load(file_obj, Loader=Loader), 'dumps': lambda obj: yaml.dumps(obj, default_flow_style=False, Dumper=Dumper)}
@@ -200,6 +200,63 @@ class Configuration(object):
 		Save the current configuration to disk.
 		"""
 		with open(self.configuration_file, 'w') as file_h:
+			file_h.write(self._serializer('dumps', self._storage))
+
+class MemoryConfiguration(Configuration):
+	"""
+	This class provides a generic object for parsing information from variables in multiple formats.
+	"""
+
+	def __init__(self, mem_object, object_type, prefix=''):
+		"""
+		:param str mem_object: The memory object to parse.
+		:param str prefix: String to be prefixed to all option names.
+		:param str object_type: String to identify type of serializer to use.
+		"""
+		self.prefix= prefix
+		self.seperator = '.'
+		self.object_type = object_type
+		if type(mem_object) != 'dict':
+			self._storage = dict(self._serializer('loads', mem_object))
+		else:
+			self._storage = mem_object
+
+	def _serializer(self, operation, *args):
+		if not self.object_type in SERIALIZER_DRIVERS:
+			raise ValueError('unknown type \'' + self.object_type + '\'')
+		function = SERIALIZER_DRIVERS[self.object_type][operation]
+		return function(*args)
+
+
+	def get_missing(self, mem_object, object_type):
+		"""
+		Use a verification configuration which has a list of required options
+		and their respective types. This information is used to identify missing
+		and incompatible options in the loaded configuration.
+
+		:param str mem_object: The file to load for verification data.
+		:param str object_type: Type to parse the mem_object as.
+		:return: A dictionary of missing and incompatible settings.
+		:rtype: dict
+		"""
+		vconf = MemoryConfiguration(mem_object, object_type)
+		missing = {}
+		for setting, setting_type in vconf.get('settings').items():
+			if not self.has_option(setting):
+				missing['missing'] = missing.get('settings', [])
+				missing['missing'].append(setting)
+			elif not type(self.get(setting)).__name__ == setting_type:
+				missing['incompatible'] = missing.get('incompatible', [])
+				missing['incompatible'].append((setting, setting_type))
+		return missing
+
+	def save(self, filename):
+		"""
+		Save the current configuration to disk.
+
+		:param str filename: The string of the file of which to save configuration to.
+		"""
+		with open(filename, 'w') as file_h:
 			file_h.write(self._serializer('dumps', self._storage))
 
 def main():
